backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic==2.7.3
weaviate-client==4.6.4
langchain==0.2.7
langchain-openai==0.1.16
langchain-text-splitters==0.2.2
langgraph==0.0.46
python-dotenv==1.0.1
arize-phoenix==3.20.0
